fix: 스페이스 탈퇴 로직 변경
스페이스 탈퇴 유효성 검사 추가
스페이스 탈퇴 후 유저가 존재하지 않으면 스페이스부터 모든 하위 노드까지 isDeleted = true

+++ note/src/main/java/com/sixb/note/api/service/SpaceService.java
@@ -184,19 +184,26 @@ public class SpaceService {
     }
 
     // 스페이스 탈퇴
-    public void leaveSpace(long userId, String spaceId) throws NotFoundException {
-        List<Space> userSpaces = spaceRepository.findSpaces(userId);
-        Space spaceToLeave = userSpaces.stream()
-                .filter(space -> space.getSpaceId().equals(spaceId))
-                .findFirst()
-                .orElseThrow(() -> new NotFoundException("스페이스를 찾을 수 없습니다."));
+    public void leaveSpace(long userId, String spaceId) throws SpaceNotFoundException {
+        Space space = spaceRepository.findSpaceById(spaceId);
 
+        if (space == null) {
+            throw new SpaceNotFoundException("존재하지 않는 스페이스입니다.");
+        }
+
+        if (!spaceRepository.isJoinedUser(userId, spaceId)) {
+            throw new SpaceNotFoundException("스페이스에 가입되지 않았습니다.");
+        }
+
+        if (!spaceRepository.isPublicSpace(spaceId)) {
+            throw new SpaceNotFoundException("개인 스페이스는 탈퇴할 수 없습니다.");
+        }
 
         spaceRepository.removeUserFromSpace(userId, spaceId);
 
         List<User> remainingUsers = userRepository.findUsersBySpaceId(spaceId);
         if (remainingUsers.isEmpty()) {
-            spaceRepository.deleteById(spaceId);
+            spaceRepository.deleteSpace(spaceId);
         }
     }
